9-18-21 A - Added and Connected GUI
Wrote GUI frontend for the project in PySide2 and connected the byte string logic to each of the widgets.  Moving sliders and clicking the animation buttons calculates the hex string to be sent to the lights.  No bleak connection yet (so can't actually connect to the lights _yet_, but getting there!