Fixed Spelling
I have been saying "Neweer" instead of "Neewer" - so I fixed the main branch's spellings.  I have a working prototype for Python that does successfully, but not really satisfied with the logistics yet.  Working on implementing a "client/server" system with zmq to more properly implement the asyncio needed for bleak.