9-22-21 - Working Prototype!
This commit fixes the issue where the bytes were reversed for CCT bytestring calculation (I had brightness and color temperature reversed), and adds integration with bleak.  It's not 100% of the way that I want it yet, PySide/Qt5 doesn't play well with asyncio and bleak _needs_ asyncio, so I need to figure out a better way of implementing them, but I do have it so that it works for testing - loading the bleak calls freezes the main GUI for a short period of time, but the scan completes, aggregates the list of Neewer lights and lets you send commands to those lights.  Also fixed "Neewer" (as I was calling it "Neweer" the whole time...)