10-11-21 - More work on command-line interpreter
Command-line interpreter works! ...when Bleak works.  Still working out the kinks, but if you launch NeewerLite-Python with command-line switches, like:

--light=F2:55:41:56:1D:A7 --mode=HSL --hue=240 --sat=80 --brightness=40 --cli

The CLI (not the GUI) will send a HSL signal of 240, saturation of 80, and brightness of 40 to light F2:55:41:56:1D:A7 - but Bleak has been flaky in testing, so most of the time it *does* work, but when it doesn't connect to the light initially, it tends to spin trying to connect.  So, *works*, but not 100% rock-solid yet.