11-4-21 - Start of HTTP Server (NOT READY FOR PRIMETIME)
NOT READY FOR PRIMETIME - just posting the latest update to the HTTP server system.  Started writing the HTTP handler, changed the processing of arguments to accept either the command line, or the HTTP parser.  Added a few more options to the argument parser to be more forgiving of mistakes (like parameters with no -- prefix, parameters with only one - for the prefix, etc.)