11-29-21 - HSL becomes HSI
I accidentally (since the beginning of this project) had "HSI" listed as "HSL" - I noticed it a while ago, but it was down the list of priorities to fix.  Now it's fixed :)  Any old mention of HSL is now HSI, the GUI is fixed, and the CLI option --mode=HSL is now --mode=HSI - also the --luminance command line parameter becomes --intensity instead.