1-16-22 - More parallelization improvements
Like the title says, more parallelization improvements!  Connecting and disconnecting is handled all at the same time, instead of one at a time, and the HTTP server now uses those techniques as well as the GUI version.

Also, as a solution to #20, NeewerLite-Python now creates a lockfile at launch, and deletes it at exit, to make sure only one copy of NEP is running at one time.  This could be refined a bit, into a "do you want to launch a new copy anyway" prompt, but that's not part of it *yet*.