2-9-22 - Store power state in snapshot presets
Added the ability to store the power state of lights in snapshot presets where they're turned off.  If a light is set to "off", it will be recorded in the preset, and if another preset needs the light to be on, it will turn on during recall.  This *should* fix the last comment of #23.