11-19-22 - Added Infinity Support
Added support for the newer Neewer light Infinity Bluetooth command protocol.  The new protocol is similar, but different in that it sends the entire MAC address of the light as part of the message being sent.

WIP: need to add G/M compensation to the CCT tab
WIP: need to work on getting proper MAC addresses on MacOS as GUIDs will not work with Infinity
WIP: ...a *LOT* with animation modes - they need a lot more than the typical "Scene" mode of the older lights!